% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{regression_plot}
\alias{regression_plot}
\title{Create a scatterplot with the added regression function}
\usage{
regression_plot(
  mod,
  data,
  colors = c("#ff6c59", "#ffd859"),
  showscale = FALSE,
  showlegend = FALSE,
  n = 50
)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Create a scatterplot with the added regression function
}
