% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\name{surface}
\alias{surface}
\title{Create a surface plot for any function}
\usage{
surface(
  expr,
  xlim = c(0, 20),
  ylim = c(0, 20),
  colors = c("#716ca8", "#6ca6a8"),
  showscale = FALSE,
  showlegend = FALSE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  n = 100
)
}
\arguments{
\item{zlab}{}
}
\value{

}
\description{
Create a surface plot for any function
}
