% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{regression_plot}
\alias{regression_plot}
\title{Title}
\usage{
regression_plot(
  mod,
  xvar,
  yvar,
  zvar,
  data,
  colors = c("#716ca8", "#6ca6a8"),
  showscale = FALSE,
  showlegend = FALSE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  n = 100
)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
