% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simply_scatter.R
\name{simply_scatter}
\alias{simply_scatter}
\title{Title}
\usage{
simply_scatter(
  x,
  y,
  z,
  col = rep(1, length(x)),
  colors = c("#000000", "red"),
  xlab = "x",
  ylab = "y",
  zlab = "z"
)
}
\arguments{
\item{zlab}{}
}
\value{

}
\description{
Title
}
