% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\name{surface}
\alias{surface}
\title{Create a 3D surface plot for any function}
\usage{
surface(
  expr,
  xlim = c(0, 40),
  ylim = c(0, 40),
  colors = c("#ff6c59", "#ffd859"),
  showscale = FALSE,
  showlegend = FALSE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  n = 100
)
}
\arguments{
\item{expr}{An expression or a function containing x and y as variables}

\item{xlim}{A vector of limits for the x-axis}

\item{ylim}{A vector of limits for the y-axis}

\item{colors}{A vector of two colors (can also be the same)}

\item{showscale}{Option for showing the colorscale}

\item{showlegend}{Option for showing the legend}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{zlab}{y-axis label}

\item{n}{number of points where function is evalutated (actually n^2)}
}
\value{
Returns a interactive 3D surface plot
}
\description{
Create a 3D surface plot for any function
}
